+++ llvm/lib/CodeGen/AsmPrinter/CodeViewDebug.cpp
@@ -1550,7 +1550,8 @@ void CodeViewDebug::beginFunctionImpl(const MachineFunction *MF) {
         GV.hasFnAttribute(Attribute::StackProtectReq)) {
       FPO |= FrameProcedureOptions::StrictSecurityChecks;
     }
-  } else if (!GV.hasStackProtectorFnAttr()) {
+  } else if (!GV.hasStackProtectorFnAttr() ||
+             GV.hasFnAttribute("stack-protector-layout-only")) {
     // __declspec(safebuffers) disables stack guards.
     FPO |= FrameProcedureOptions::SafeBuffers;
   }