+++ llvm/lib/CodeGen/StackProtector.cpp
@@ -55,6 +55,8 @@ using namespace llvm;
 #define DEBUG_TYPE "stack-protector"
 
 STATISTIC(NumFunProtected, "Number of functions protected");
+STATISTIC(NumFunLayoutOnly,
+          "Number of functions with only the stack protector layout applied");
 STATISTIC(NumAddrTaken, "Number of local variables that have their address"
                         " taken.");
 
@@ -100,7 +102,21 @@ void SSPLayoutInfo::copyToMachineFrameInfo(MachineFrameInfo &MFI) const {
       continue;
 
     MFI.setObjectSSPLayout(I, LI->second);
+    if (LayoutOnly)
+      MFI.setStackProtectorLayoutOnly();
   }
+}
+
+/// Return true if only the stack protector layout is applied to \p F and no
+/// stack protector is inserted. Functions with funclet-based personalities are
+/// not supported by the stack protector, so they get neither.
+static bool isLayoutOnly(const Function &F) {
+  if (!F.hasFnAttribute("stack-protector-layout-only"))
+    return false;
+  if (F.hasPersonalityFn() &&
+      isFuncletEHPersonality(classifyEHPersonality(F.getPersonalityFn())))
+    return false;
+  return true;
 }
 
 SSPLayoutInfo SSPLayoutAnalysis::run(Function &F,
@@ -109,6 +125,7 @@ SSPLayoutInfo SSPLayoutAnalysis::run(Function &F,
   SSPLayoutInfo Info;
   Info.RequireStackProtector =
       SSPLayoutAnalysis::requiresStackProtector(&F, &Info.Layout);
+  Info.LayoutOnly = isLayoutOnly(F);
   Info.SSPBufferSize = F.getFnAttributeAsParsedInteger(
       "stack-protector-buffer-size", SSPLayoutInfo::DefaultSSPBufferSize);
   return Info;
@@ -124,6 +141,11 @@ PreservedAnalyses StackProtectorPass::run(Function &F,
 
   if (!Info.RequireStackProtector)
     return PreservedAnalyses::all();
+
+  if (Info.LayoutOnly) {
+    ++NumFunLayoutOnly;
+    return PreservedAnalyses::all();
+  }
 
   // TODO(etienneb): Functions with funclets are not correctly supported now.
   // Do nothing if this is funclet-based personality.
@@ -191,6 +213,8 @@ bool StackProtector::runOnFunction(Function &Fn) {
   if (auto *DTWP = getAnalysisIfAvailable<DominatorTreeWrapperPass>())
     DTU.emplace(DTWP->getDomTree(), DomTreeUpdater::UpdateStrategy::Lazy);
   TM = &getAnalysis<TargetPassConfig>().getTM<TargetMachine>();
+  LayoutInfo.Layout.clear();
+  LayoutInfo.LayoutOnly = isLayoutOnly(Fn);
   LayoutInfo.HasPrologue = false;
   LayoutInfo.HasIRCheck = false;
 
@@ -198,6 +222,11 @@ bool StackProtector::runOnFunction(Function &Fn) {
       "stack-protector-buffer-size", SSPLayoutInfo::DefaultSSPBufferSize);
   if (!requiresStackProtector(F, &LayoutInfo.Layout))
     return false;
+
+  if (LayoutInfo.LayoutOnly) {
+    ++NumFunLayoutOnly;
+    return false;
+  }
 
   // TODO(etienneb): Functions with funclets are not correctly supported now.
   // Do nothing if this is funclet-based personality.